[SA-50909] Do not show the alert consent before SSO login on <iOS 13 (#2)
We got rid of the “annoying” prompt before showing the SSO login page on iOS 13+ by using the system-provided ephemeral browser session. Prior to iOS 13, it is not an option.

Reading up on threads ([AppAuth-iOS#402](https://github.com/openid/AppAuth-iOS/issues/402)), showcasing a similar situation to ours, it seems like the only solution for us is to use the `SFSafariViewController` for authentication on all devices below iOS 13.