[SA-55649] Fix crash on SFSafariViewController fallback
Problem:
`ASWebAuthenticationSession.start()` can return `false` meaning that it failed to start, but that automatically calls the `completionHandler` even before returning, which will invalidate the session automatically. In this case, the code falls back to `SFSafariViewController` which even on a successful login will fail and crash the application because the SDK will try to process an authorization response when the session has already been invalidated.

Resolution:
Make sure to only process completionHandler from `ASWebAuthenticationSession` when the auth presentation was successful.